SpreadsheetListComponentSpreadsheetDataTableComponentCellRenderer extracted
